feat: move merged observable implementation into the observable file
